+++ CHANGELOG.md
@@ -15,6 +15,7 @@ No changes to highlight.
 ## Other Changes:
 
 - Fix Faster R-CNN detection head by `@illian01` in [PR 184](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/184)
+- Refactoring models/op module by `@illian01` in [PR 189](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/189)
 
 # v0.0.8